Moved the java files to one directory
